+++ src/main/kotlin/io/runnershigh/backend/training/domain/mapper/TrainingScheduleMapper.kt
@@ -1,7 +1,7 @@
 package io.runnershigh.backend.training.domain.mapper
 
 import io.runnershigh.backend.training.infrastructure.entity.TrainingSchedules
-import io.runnershigh.backend.training.ui.dto.SaveTrainingSchedule
+import io.runnershigh.backend.training.ui.dto.request.SaveTrainingSchedule
 import io.runnershigh.backend.user.infrastructure.entity.UserEntity
 
 // Training schedule mapper
